requirements.txt
MyProxyClient>=2.0.1
pyOpenSSL>=17.0.0
